+++ oriedita-data/src/main/java/oriedita/editor/export/api/FileImporter.java
@@ -1,4 +1,4 @@
-package oriedita.filesupport.api;
+package oriedita.editor.export.api;
 
 import oriedita.editor.save.Save;
